Closes #19. Move some exceptions to single file.
