+++ src/exceptions/exceptions.ts
@@ -1,0 +1,10 @@
+export class InvalidParametersException extends Error {
+  constructor(public details: any) {
+    super(details);
+    this.name = 'Invalid request';
+  }
+}
+
+export class NotFoundException extends Error {
+
+}
