--- src/exceptions/invalid.parameters.ts
@@ -1,6 +1,0 @@
-export default class InvalidParametersException extends Error {
-  constructor(public details: any) {
-    super(details);
-    this.name = 'Invalid request';
-  }
-}
