--- src/exceptions/not.found.ts
@@ -1,3 +1,0 @@
-export default class NotFoundException extends Error {
-
-}
